This commit to fix the final package names and keep Andreas Reichel as the only developer of this initial version.
