+++ src/test/java/ai/starlake/transpiler/DebugTest.java
@@ -1,21 +1,4 @@
-/**
- * Manticore Projects JSQLTranspiler is a multiple SQL Dialect to DuckDB Translation Software.
- * Copyright (C) 2024 Andreas Reichel <andreas@manticore-projects.com>
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published
- * by the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- */
-package com.manticore.transpiler;
+package ai.starlake.transpiler;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
@@ -28,7 +11,7 @@ import java.util.stream.Stream;
 // The purpose of this facility is to debug one single test line by line
 // Since this is not easy when using the parametrised tests
 public class DebugTest extends JSQLTranspilerTest {
-  public final static String TEST_FOLDER_STR = "build/resources/test/com/manticore/transpiler/any";
+  public final static String TEST_FOLDER_STR = "build/resources/test/ai/starlake/transpiler/any";
 
   public static final FilenameFilter FILENAME_FILTER = new FilenameFilter() {
     @Override
