+++ src/test/java/ai/starlake/transpiler/JSQLTranspilerTest.java
@@ -1,21 +1,4 @@
-/**
- * Manticore Projects JSQLTranspiler is a multiple SQL Dialect to DuckDB Translation Software.
- * Copyright (C) 2024 Andreas Reichel <andreas@manticore-projects.com>
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published
- * by the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- */
-package com.manticore.transpiler;
+package ai.starlake.transpiler;
 
 import com.opencsv.CSVWriter;
 import com.opencsv.ResultSetHelperService;
@@ -82,7 +65,7 @@ class JSQLTranspilerTest {
   private static final Pattern SQL_SANITATION_PATTERN2 =
       Pattern.compile("\\s*([!/,()=+\\-*|\\{\\}\\[\\]<>:])\\s*", Pattern.MULTILINE);
 
-  public final static String TEST_FOLDER_STR = "build/resources/test/com/manticore/transpiler/any";
+  public final static String TEST_FOLDER_STR = "build/resources/test/ai/starlake/transpiler/any";
 
   public static final FilenameFilter FILENAME_FILTER = new FilenameFilter() {
     @Override
@@ -244,7 +227,7 @@ class JSQLTranspilerTest {
       // Download the ZIP file via Gradle Download task in order to enable caching
       // URL url = new URL("https://docs.aws.amazon.com/redshift/latest/gsg/samples/tickitdb.zip");
       URL url = JSQLTranspilerTest.class.getClassLoader()
-          .getResource("com/manticore/transpiler/tickitdb.zip");
+          .getResource("ai/starlake/transpiler/tickitdb.zip");
       assert url != null;
       try (InputStream in = url.openStream()) {
         // Extract the ZIP file
