+++ src/test/resources/ai/starlake/transpiler/JSQLTranspilerTest_DDL.sql
