+++ src/test/resources/ai/starlake/transpiler/JSQLTranspilerTest_MainIn.sql
