+++ src/test/resources/ai/starlake/transpiler/JSQLTranspilerTest_MainOut.sql
