+++ src/test/resources/ai/starlake/transpiler/amazon_redshift/select_top_clause_nested.sql
