+++ src/test/resources/ai/starlake/transpiler/any/debug.sql
