+++ src/test/resources/ai/starlake/transpiler/any/nvl_function.sql
