+++ src/test/resources/ai/starlake/transpiler/any/select_top_clause.sql
