+++ src/test/resources/ai/starlake/transpiler/google_bigquery/aggregate_function.sql
