+++ src/test/resources/ai/starlake/transpiler/google_bigquery/approx_aggregate_function.sql
