+++ src/test/resources/ai/starlake/transpiler/google_bigquery/array_functions.sql
