+++ src/test/resources/ai/starlake/transpiler/google_bigquery/date_add_function.sql
