+++ src/test/resources/ai/starlake/transpiler/google_bigquery/date_format.sql
