+++ src/test/resources/ai/starlake/transpiler/google_bigquery/date_function.sql
