+++ src/test/resources/ai/starlake/transpiler/google_bigquery/date_trunc_function.sql
