+++ src/test/resources/ai/starlake/transpiler/google_bigquery/datetime_function.sql
