+++ src/test/resources/ai/starlake/transpiler/google_bigquery/disabled.sql
