+++ src/test/resources/ai/starlake/transpiler/google_bigquery/extract_function.sql
