+++ src/test/resources/ai/starlake/transpiler/google_bigquery/last_day_function.sql
