+++ src/test/resources/ai/starlake/transpiler/google_bigquery/math_functions.sql
