+++ src/test/resources/ai/starlake/transpiler/google_bigquery/navigation.sql
