+++ src/test/resources/ai/starlake/transpiler/google_bigquery/number_or_rank_function.sql
