+++ src/test/resources/ai/starlake/transpiler/google_bigquery/parse_date_function.sql
