+++ src/test/resources/ai/starlake/transpiler/google_bigquery/structure.sql
