+++ src/test/resources/ai/starlake/transpiler/google_bigquery/text_function.sql
