+++ src/test/resources/ai/starlake/transpiler/google_bigquery/time_function.sql
