+++ src/test/resources/ai/starlake/transpiler/google_bigquery/timestamp_function.sql
