+++ src/test/resources/ai/starlake/transpiler/google_bigquery/unix_date_function.sql
