+++ src/test/resources/ai/starlake/transpiler/google_bigquery/utility_function.sql
